requirements.txt
Flask==0.9
PasteDeploy==1.5.0
